Folder stucture User services added

+++ src/main/java/com/ETechSustain/ETechSustain/Repository/UsuariosRepository.java
@@ -4,7 +4,10 @@ import com.ETechSustain.ETechSustain.Entity.Usuarios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+import java.util.Optional;
+
 @Repository
-public interface UsuariosRepository extends JpaRepository<Usuarios, Long> {
+public interface UsuariosRepository extends JpaRepository<Usuarios, Integer> {
 
 }
